// src/entities/Profile/model/types/profile.ts
import { Currency } from 'entities/Currency/model/types/currency';
import { Country } from 'entities/Country/model/types/country';

export enum ValidateProfileError {
	NO_DATA = 'NO_DATA',
	SERVER_ERROR = 'SERVER_ERROR',
	INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
	INCORRECT_AGE = 'INCORRECT_AGE',
	INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
	INCORRECT_CITY = 'INCORRECT_CITY',
	INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
	INCORRECT_USERNAME = 'INCORRECT_USERNAME',
	INCORRECT_AVATAR = 'INCORRECT_AVATAR',
}

export interface Profile {
	'first'?: string;
	'lastname'?: string;
	'age'?: number;
	'currency'?: Currency;
	'country'?: Country;
	'city'?: string;
	'username'?: string;
	'avatar'?: string;
}

export interface ProfileSchema {
	data?: Profile;
	form?: Profile;
	isLoading: boolean;
	error?: string;
	readonly: boolean;
	validateErrors?: ValidateProfileError[];
}